Slurm

+++ script/template/mdrun.sh
@@ -13,12 +13,19 @@ fi
 export OMP_NUM_THREADS=$ncpus
 ## for thread-MPI parallelization
 
+finished_info=finished_step_list
+touch $finished_info
+
 for stepname in {{ STEP_NAMES }}
 do
     prev_cpt=""
     prev=$now
     now=$stepname
 
+    if [ `grep -x ${now} $finished_info | wc -l` = 1 ] ;then
+       continue
+    fi
+    
     rm -f ${now}.out.mdp ${now}.tpr ${now}.log ${now}.gro ${now}.trr ${now}.edr ${now}.cpt
     echo $GMX grompp -f ${now}.mdp -po ${now}.out.mdp -o ${now}.tpr \
 	-c ${prev}.gro -p ${top} \
@@ -28,7 +35,7 @@ do
     $GMX grompp -f ${now}.mdp -po ${now}.out.mdp -o ${now}.tpr \
 	  -c ${prev}.gro -p ${top} \
 	  -r ${prev}.gro -n index.ndx ${prev_cpt}
-    $GMX mdrun -reprod -v -s ${now} -deffnm ${now} -cpi ${now}.cpt
+    $GMX mdrun -reprod -v -s ${now} -deffnm ${now} -cpi ${now}.cpt && echo $now >> $finished_info
 
     prev_cpt="-t $now.cpt"
 done
